Metodo destroy de fabricante controller y Actualizacion fabricante avion controller
Metodo destroy de fabricante controller y Actualizacion fabricante avion
controller

+++ app/Http/Controllers/FabricanteController.php
@@ -104,7 +104,7 @@ class FabricanteController extends Controller {
 		//si hay datos devolvemos un codigo 200 y si no los hay devolvemos un 404
 		if (!$fabricante) {
 			//se devuelve un array errors con los errores y codigo 404
-			return response()->json(['errors' => Array(['code' => 404, 'message' => 'No se encuentra					un fabricante con ese código'])], 404);
+			return response()->json(['errors' => Array(['code' => 404, 'message' => 'No se encuentra un fabricante con ese código'])], 404);
 		}
 		//devolvemos la informacion encontrada
 		return response()->json(['status' => 'ok', 'data' => $fabricante], 200);
@@ -129,10 +129,77 @@ class FabricanteController extends Controller {
 	 * @param  int  $id
 	 * @return Response
 	 */
-	public function update($id) {//para put y patch
+	public function update($id, Request $request) {//para put y patch//PROBAR!!!!!!!!!!!!
 	// en el put se actualizan los datos y en el patch se actualiza uno de los campos y en los dos se entra x update
 	//asi q tengo q comprobar...
-		//
+		// para actualizar un fabricante se le pasa el id (en la ruta)
+		//vamos a actualizar un fabricante y para eso comprobamos si el fabricante existe o no
+		//en otro caso devolvemos error
+		$fabricante=Fabricante::find($id);
+		//si no existe mostramos error
+		if(!$fabricante)
+		{   //return response()->json(['errors' => Array(['code' => 404, 'message' => 'No se /
+			//	un fabricante con ese código'])], 404);/
+			return response()->json(['errors'=>array(['code'=>404, "message"=>"error"])], 404);
+		}
+		//almacenamos variables para facilitar el uso, los campos recibidos
+		$nombre=$request->input('nombre');
+		$direccion=$request->input('direccion');
+		$telefono=$request->input('telefono');
+		
+		//comprobamos si recibimos peticion x patch
+		if($request->method()=='PATCH')
+		{
+			$bandera=false;
+			//ACTUALIZACION PARCIAL DE DATOS
+				if($nombre !=null && $nombre!='')
+				{
+					$fabricante->nombre=$nombre;
+					$bandera=true;
+				}
+			//ACTUALIZACION PARCIAL DE DATOS
+				if($direccion !=null && $direccion!='')
+				{
+					$fabricante->direccion=$direccion;
+					$bandera=true;
+				}
+				//ACTUALIZACION PARCIAL DE DATOS
+				if($telefono !=null && $telefono!='')
+				{
+					$fabricante->telefono=$telefono;
+					$bandera=true;
+				}
+				if ($bandera)
+				{
+					//grabamos fabricante
+					$fabricante->save();
+					//devolvemos un 200
+					return response()->json(['status'=>'ok','data'=>$fabricante],200);
+				}
+				else
+				{
+					//devolvemos un 304 not modified
+					return response()->json(['errors'=>array(['code'=>304, 'message'=>"No se ha modificado ningun dato del fabricante"])],304);
+				}
+		}
+		
+		
+		//METODO PUT actualizamos todos los campos
+		//comprobamos q recibimos todos
+		if (!$nombre || !$direccion || $telefono)
+		{
+			//de devuelve codigo 422 q sig q no se pden procesar los campos
+			return response()->json(['errors'=>array(['code=>422','message'=>"Faltan valores"])],422);
+			//actualizamos los 3 campos
+			$fabricante->nombre=$nombre;
+			$fabricante->direccion=$direccion;
+			$fabricante->telefono=$telefono;
+			
+			//grabamos el fabricante
+			$fabricante->save();
+			return response()->json(['status'=>'ok','data'=>$fabricante],200);
+			
+		}
 	}
 
 	/**
@@ -154,8 +221,25 @@ class FabricanteController extends Controller {
 		//204 sig NO CONTENT
 		//ESTE CODIGO no mustra texto en el body
 		//si qsieramos ver el mensaje devolveriamos un codigo 200
+		//antes de borrarlo comprobamos si tiene aviones y si es asi 
+		//sacamos un mensaje d error
+		//$aviones = $fabricante->aviones()->get();
+		$aviones = $fabricante->aviones;		
+		if (sizeof($aviones)>0)
+		{
+			
+			//si quisieramos borrar todos los aviones del fabricante seria:
+			//$fabricante->aviones->delete();
+			//
+			
+			
+			//devolvemos un codigo de conflicto 409
+			return response()->json(['errors' => array(['code' => 409, "messagge" => "Este fabricante posee aviones y no puede ser eliminado"])], 409);
+		}
+		//eliminamos el fabricante si no tiene aviones
 		$fabricante->delete();
 
+		//se devuelve el codigo 204 No content
 		return response()->json(['code' => 204, "message" => "Se ha eliminado correctamente el fabricante"], 204);
 	}
